css: make tooltip text selectable
Total hack using ::before to create an invisible area that receives
pointer events while the tooltip is visible, since the tooltip itself is
some area above the item that needs to be hovered. This is needed to
ensure the user can easily hover over to the tooltip's text without
pointer events going away.