css, layouts: fix overflow on mobile
Tooltips can be extremely long, especially on mobile. It turns out that
most mobile web browsers ignore the overflow attributes on the body and
html elements, so we create another wrapper with this attribute and
adjust margins on the body and this wrapper so as not to clip the
tooltips should they go off screen.